% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.warm.start.R
\name{clean.warm.start}
\alias{clean.warm.start}
\title{rdname softImpute-internal}
\usage{
clean.warm.start(a)
}
\arguments{
\item{a}{an svd object with components \code{u}, \code{d} and \code{v} or \code{NULL}}
}
\description{
rdname softImpute-internal
}
