\name{softImpute-internal}
\title{Internal softImpute functions}
\alias{clean.warm.start}
\alias{simpute.als}
\alias{simpute.svd}
\alias{Ssimpute.als}
\alias{Ssimpute.svd}
\alias{Ssvd.als}
\alias{softImpute.x.matrix}
\alias{softImpute.x.Incomplete}
\description{Internal softImpute functions}
\author{Trevor Hastie}
\details{
  These functions are not intended to be called directly, but they can
  be useful for understanding the structure of the models used.
}
\keyword{internal}
